ingestion/pyproject.toml
[project]
name = "engram-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.68",
    "pika>=1.3.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.2",
    "python-magic>=0.4.27",
]
